requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
defusedxml==0.7.1
Django==5.2.5
django-debug-toolbar==6.0.0
django-filter==25.2
django-htmx==1.25.0
django-js-asset==3.1.2
django-mptt==0.18.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Markdown==3.9
oauthlib==3.3.1
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.9.0
shortuuid==1.0.13
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
urllib3==2.5.0
zope.interface==8.0.1
